% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gp1.R
\name{count_wells_from_header}
\alias{count_wells_from_header}
\title{count the number of fields that correspond to wells from file header}
\usage{
count_wells_from_header(path, n_fields_other = 3L)
}
\arguments{
\item{path}{\code{<chr>} path to raw file}

\item{n_fields_other}{\code{<int>} number of fields that are not wells. default: 3}
}
\value{
\code{<int>} number of wells (fields).
}
\description{
count the number of fields that correspond to wells from file header
}
\keyword{internal}
