% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gp1.R
\name{gp1_example}
\alias{gp1_example}
\title{get a gp1 example data file}
\usage{
gp1_example(file)
}
\arguments{
\item{file}{\code{<chr>} name of the file}
}
\description{
get a gp1 example data file
}
\examples{
    gp1_example('GP1-Results_feb11.txt')
    gp1_example('GP1-Results_jan27.txt')
    gp1_example('HIPHOP-Feb27Results.txt')
}
