% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gp1.R
\name{mtp_wells}
\alias{mtp_wells}
\title{get well labels from number of rows and cols}
\usage{
mtp_wells(nrow, ncol, order = c("rowmajor", "colmajor"))
}
\arguments{
\item{nrow}{<int> number of rows}

\item{ncol}{<int> number of cols}

\item{order}{<chr> defaults to 'rowmajor', alternatively 'colmajor'}
}
\description{
get well labels from number of rows and cols
}
