% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gp1.R
\name{read_gp1}
\alias{read_gp1}
\title{read GP-1 output file into tidy tibble}
\usage{
read_gp1(path, well_labels = gp1::mtp_wells(6, 8))
}
\arguments{
\item{path}{<chr> path to results file}

\item{well_labels}{<chr> vector of well labels, defaults to row major 48MTP ('A01', 'A02',...)}
}
\description{
read GP-1 output file into tidy tibble
}
