% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gp1.R
\name{read_gp1}
\alias{read_gp1}
\title{read GP-1 output file into tidy tibble}
\usage{
read_gp1(file, all_fields = FALSE)
}
\arguments{
\item{file}{\code{<chr>} path to results file}

\item{all_fields}{\code{<lgl>} should all fields be included? Defaults to \link{FALSE}.}
}
\value{
\code{\link[=tibble]{tibble()}} with the following cols:
\itemize{
\item \strong{plate} identifier as found in file,
\item \strong{well} identifier,
\item \strong{runtime} time since run start in seconds,
\item \strong{measure} numeric value of measurement as recorded,
}

if \code{all_fields = TRUE}:
\itemize{
\item \strong{datetime} of measurement in ISO8601 (datetime format),
\item \strong{measure_type} the type of measurement; currently this is a stub to
allow for multi-measurement-type results parsing in the future if needed.
}
}
\description{
read GP-1 output file into tidy tibble
}
\examples{
file <- gp1_example('GP1-Results_feb11.txt')
read_gp1(file)
read_gp1(file, all_fields = TRUE)
}
